requirements.txt
Django==1.10.2
django-haystack==2.5.0
django-polls==0.1
elasticsearch==2.4.0
urllib3==1.18
